HAPP-1600 Removed un-used code - deleted un-used code for dependents
