--- app/src/main/res/layout/item_dependent_aged_out.xml
@@ -1,81 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_marginBottom="@dimen/space_2_x"
-    android:background="@drawable/layout_shape_round"
-    android:filterTouchesWhenObscured="true"
-    android:padding="@dimen/space_2_x"
-    android:paddingStart="@dimen/space_2_x"
-    android:paddingTop="@dimen/space_1_x"
-    android:paddingEnd="@dimen/space_2_x"
-    android:paddingBottom="@dimen/space_3_x">
-
-    <ImageView
-        android:id="@+id/iv_icon"
-        android:layout_width="wrap_content"
-        android:layout_height="26dp"
-        android:src="@drawable/ic_manage_dependent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:ignore="ContentDescription" />
-
-    <com.google.android.material.textview.MaterialTextView
-        android:id="@+id/txt_label"
-        style="?textAppearanceBody2"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="12dp"
-        android:gravity="center_vertical"
-        android:textSize="16sp"
-        android:textStyle="bold"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/iv_icon"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:text="John Tester" />
-
-    <LinearLayout
-        android:id="@+id/view_warning"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="@dimen/space_1_x"
-        android:background="@drawable/shape_rounded_rectangle_warning"
-        android:gravity="top"
-        android:orientation="horizontal"
-        android:padding="@dimen/space_2_x"
-        app:layout_constraintEnd_toEndOf="@id/txt_label"
-        app:layout_constraintStart_toStartOf="@+id/iv_icon"
-        app:layout_constraintTop_toBottomOf="@+id/txt_label">
-
-        <ImageView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:src="@drawable/ic_warning" />
-
-        <com.google.android.material.textview.MaterialTextView
-            android:id="@+id/txt_body"
-            style="?textAppearanceBody2"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="@dimen/space_1_x"
-            android:layout_marginBottom="@dimen/space_1_x"
-            android:text="@string/dependents_aged_out_message"
-            android:textColor="@color/warning_text"
-            android:textSize="16sp" />
-
-    </LinearLayout>
-
-    <com.google.android.material.button.MaterialButton
-        android:id="@+id/btn_remove"
-        style="@style/HealthGateway.Widget.MaterialComponents.Button.OutlinedButton"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="@dimen/space_1_x"
-        android:filterTouchesWhenObscured="true"
-        android:text="@string/dependent_remove"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/view_warning" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
